Revert "Added makefile to T3 assignments Materijali kuhinje, Subjekat-predika…"
This reverts commit 229321abdc795b1984ced7d712a452c39a6e7d65.
