+++ zadaci/2023/sov/subjekat-predikat-objekat/resenje.c
@@ -2,14 +2,28 @@
 #include <string.h>
 #include <ctype.h>
 #include <stdlib.h>
-#include "lista.h"
+
+#define MAX_REC 21
+#define MAX_VRSTA 9
 
 #define MAX_NAZIV_IZLAZNE 11
+
+typedef struct rec_st
+{
+    char rec[MAX_REC];
+    char vrsta[MAX_VRSTA];
+    struct rec_st *sledeci;
+} REC;
 
 FILE *safe_fopen(char *, char *, int);
 void ucitaj_reci(FILE *, REC **);
 void ispisi_recenice(FILE *, REC *);
+
+void inicijalizacija(REC **);
+REC *napravi_cvor(char *, char *);
+void dodaj_na_kraj(REC **, REC *);
 int prebroj_vrstu(REC *, char *);
+void obrisi_listu(REC **);
 
 int main(int argc, char **argv)
 {
@@ -114,6 +128,46 @@ void ispisi_recenice(FILE *izlazna, REC *glava)
     }
 }
 
+void inicijalizacija(REC **pglava)
+{
+    *pglava = NULL;
+}
+
+REC *napravi_cvor(char *rec, char *vrsta)
+{
+    REC *novi = (REC *)malloc(sizeof(REC));
+
+    if(novi == NULL)
+    {
+        printf("Greska prilikom zauzimanja memorije!\n");
+        exit(2);
+    }
+
+    strcpy(novi->rec, rec);
+    strcpy(novi->vrsta, vrsta);
+    novi->sledeci = NULL;
+
+    return novi;
+}
+
+void dodaj_na_kraj(REC **pglava, REC *novi)
+{
+    if(*pglava == NULL)
+    {
+        *pglava = novi;
+    }
+    else
+    {
+        REC *tekuci = *pglava;
+
+        while(tekuci->sledeci != NULL)
+        {
+            tekuci = tekuci->sledeci;
+        }
+
+        tekuci->sledeci = novi;
+    }
+}
 
 int prebroj_vrstu(REC *glava, char *vrsta)
 {
@@ -131,4 +185,17 @@ int prebroj_vrstu(REC *glava, char *vrsta)
     }
 
     return broj;
+}
+
+void obrisi_listu(REC **pglava)
+{
+    REC *tmp;
+
+    while(*pglava != NULL)
+    {
+        tmp = *pglava;
+        *pglava = (*pglava)->sledeci;
+        tmp->sledeci = NULL;
+        free(tmp);
+    }
 }
