+++ zadaci/2024/sov/materijali-kuhinje/index.rst
@@ -13,20 +13,8 @@ Materijali kuhinje
 Primer rešenja
 --------------
 
-.. tabs::
-
-  .. group-tab:: resenje.c
-    .. literalinclude:: resenje.c 
-       :linenos:
-  .. group-tab:: lista.h
-    .. literalinclude:: lista.h 
-      :linenos:
-  .. group-tab:: lista.c
-    .. literalinclude:: lista.c 
-      :linenos:
-  .. group-tab:: Makefile
-    .. literalinclude:: Makefile
-      :linenos:
+.. literalinclude:: resenje.c 
+   :linenos:
 
 .. raw:: html
