+++ zadaci/2024/sov/materijali-kuhinje/resenje.c
@@ -2,7 +2,17 @@
 #include <stdlib.h>
 #include <string.h>
 #include <math.h>
-#include "lista.h"
+
+#define MAX_KATEGORIJA 21
+#define MAX_NAZIV 21
+
+typedef struct cvor_st{
+    char kategorija[MAX_KATEGORIJA];
+    char naziv[MAX_NAZIV];
+    float cena;
+    float ukupna_cena;
+    struct cvor_st* sledeci;
+}CVOR;
 
 FILE* open_file(char* ime, char* mode) {
 
@@ -13,6 +23,44 @@ FILE* open_file(char* ime, char* mode) {
         exit(2);
     }
     return fp;
+}
+
+void init(CVOR** glava) {
+    *glava = NULL;
+}
+
+CVOR* create_node(char* kategorija, char* naziv, float cena, float povrsina) {
+    CVOR* tmp = malloc(sizeof(CVOR));
+    if(tmp==NULL)
+    {
+        printf("Ne moze se zauzeti memorija\n");
+        exit(3);
+    }
+    strcpy(tmp->kategorija, kategorija);
+    strcpy(tmp->naziv, naziv);
+    tmp->cena = cena;
+    tmp->ukupna_cena = cena * povrsina;
+    tmp->sledeci = NULL;
+    return tmp; 
+}
+
+void add_sorted(CVOR** glava, CVOR* novi) {
+
+   CVOR* trenutni;
+    if(*glava==NULL || (*glava)->cena >= novi->cena) 
+    {
+        novi->sledeci = *glava;
+        *glava = novi;
+    } else 
+    {
+        trenutni = *glava;
+        while(trenutni->sledeci!=NULL && trenutni->sledeci->cena < novi->cena)
+        {
+            trenutni = trenutni->sledeci;
+        }
+        novi->sledeci = trenutni->sledeci;
+        trenutni->sledeci = novi;
+    }
 }
 
 void read(FILE* in, CVOR** glava, float povrsina) {
@@ -59,6 +107,16 @@ CVOR* nadji_najblizi(CVOR* glava, char* kategorija, float zeljeno){
     return najblizi;
 }
 
+void clear(CVOR** glava) {
+    CVOR *tmp;
+    while(*glava != NULL)
+    {
+        tmp = *glava;
+        *glava = (*glava)->sledeci;
+        tmp->sledeci = NULL;
+        free(tmp);
+    }
+}
 
 int main(int argc, char* argv[]) {
